requirements.txt
# Core agent stack
langchain>=0.3.14
langchain-core>=0.3.14
langchain-openai>=0.2.8
langgraph>=0.2.55
langchain-mongodb>=0.2.0
langgraph-checkpoint-mongodb>=0.1.0
# langgraph-checkpoint-mongodb currently expects pymongo<4.16
pymongo>=4.12,<4.16
certifi>=2024.0.0

# Embeddings (Voyage AI official client)
voyageai>=0.3.0

# App & utilities
python-dotenv>=1.0.1
streamlit>=1.39.0
httpx>=0.27.0
deepgram-sdk>=6.0.0
streamlit-webrtc>=0.47.0
av>=12.0.0

# Tracing (Phoenix 4+ uses OpenInference for LangChain; see agent.py)
arize-phoenix>=4.0.0
openinference-instrumentation-langchain>=0.1.0
